Legg til opplysningspåminnelse felt i alle oppfølgningspørsmål
