--- schemas/steg/om-barnet/oppfolgning/fosterbarn.ts
@@ -1,33 +1,0 @@
-import {
-  CustomSanityTyper,
-  DokumentNavn,
-  dokumentTittel,
-  SanityTyper,
-  Steg,
-  StegDokument,
-} from '../../../typer';
-
-const fosterbarn: StegDokument = {
-  steg: Steg.OM_BARNET,
-  title: dokumentTittel.OM_BARNET_FOSTERBARN,
-  name: DokumentNavn.OM_BARNET_FOSTERBARN,
-  type: SanityTyper.DOCUMENT,
-  fields: [
-    {
-      name: 'sporsmal',
-      title: 'Spørsmål',
-      description: '(obligatorisk)',
-      type: CustomSanityTyper.LOCALE_STRING,
-      validation: Rule => Rule.required().error('Du må fylle inn spørsmålsteksten'),
-    },
-    {
-      name: 'feilmelding',
-      title: 'Feilmelding',
-      description: '(obligatorisk)',
-      type: CustomSanityTyper.LOCALE_STRING,
-      validation: Rule => Rule.required().error('Du må fylle inn feilmelding'),
-    },
-  ],
-};
-
-export default fosterbarn;
